# [`Optimize`]: HashSetCache Implementation