--- a/src/Neo.IO/Caching/HashSetCache.cs
+++ b/src/Neo.IO/Caching/HashSetCache.cs
@@ -30,7 +30,6 @@ namespace Neo.IO.Caching
         }
 
         private readonly int _capacity;
-
         private readonly Items _items;
 
         /// <summary>