About this PR:
1. A better performance implementation;
2. Rename `Add` to `TryAdd` to match it's behaivor;
3. Make some unit tests clearer;
4. Add some "{ }" to make code clearer;

The current `HashSetCache` was introduced by #1389 to fix performance issues with `FIFOSet`,  and this PR submits a better implementation:
For performance test on #1389:
After(Note: run on different platforms.):
```
Add timespan: 0.0268179s
except with timespan: 8.6E-05s
contains with timespan: 6E-07s result: False
contains with timespan: 0s result: True
```


For Benchmark:

Before:
```
| Method       | OperationCount | Mean      | Error    | StdDev   |
|------------- |--------------- |----------:|---------:|---------:|
| HashSetCache | 1000           |  30.25 us | 0.186 us | 0.174 us |
| HashSetCache | 10000          | 454.59 us | 2.337 us | 2.186 us |
```

After:
```
| Method       | OperationCount | Mean      | Error    | StdDev   |
|------------- |--------------- |----------:|---------:|---------:|
| HashSetCache | 1000           |  22.81 us | 0.227 us | 0.212 us |
| HashSetCache | 10000          | 245.61 us | 0.672 us | 0.561 us |
```

## Type of change

<!-- Please delete options that are not relevant. -->

- [x] Optimization (the change is only an optimization)
- [x] Style (the change is only a code style for better maintenance or standard purpose)
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update


# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my code
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules